+++ src/main/java/com/ccl/grandcanyon/Admins.java
@@ -384,8 +384,8 @@ public class Admins {
       // send email to admin with link
       // TODO: replace this message body with HTML email template.
       ReminderService reminderService = ReminderService.getInstance();
-
-      String resetUrl = reminderService.getAdminApplicationBaseUrl() + "/finish_password_reset?token=" + token;
+      ReminderMessageFormatter RMF = ReminderMessageFormatter.getInstance();
+      String resetUrl = RMF.getAdminApplicationBaseURL() + "/finish_password_reset?token=" + token;
       Message resetRequestMessage = new Message();
       resetRequestMessage.setSubject("Password Reset Requested");
       resetRequestMessage.setBody("If you requsted a password reset, visit the page at the following URL:  http://" + resetUrl +
