+++ src/main/java/com/ccl/grandcanyon/types/Message.java
@@ -1,9 +1,14 @@
 package com.ccl.grandcanyon.types;
 
+import com.ccl.grandcanyon.types.District;
+
+import java.util.List;
+
 public class Message {
 
   private String subject;
   private String body;
+  private List<District> targets;
 
   public String getSubject() {
     return subject;
@@ -19,5 +24,13 @@ public class Message {
 
   public void setBody(String body) {
     this.body = body;
+  }
+
+  public List<District> getTargetDistricts() {
+    return targets;
+  }
+  
+  public void addTargetDistrict(District targetDistrict) {
+    this.targets.add(targetDistrict);
   }
 }
